training/lc_classifiers/models/experimentation/ATAT/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorlog==6.9.0
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.42.0
Deprecated==1.2.18
docker==7.1.0
filelock==3.17.0
Flask==3.1.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.11.9
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mlflow==2.20.1
mlflow-skinny==2.20.1
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.2.1
protobuf==5.29.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
scikit-learn==1.4.2
scipy==1.15.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.37
sqlparse==0.5.3
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
torch==2.3.1
torchmetrics==1.6.1
tqdm==4.67.1
triton==2.3.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
